Fixing type

+++ src/jsd_epd_nominal.c
@@ -51,7 +51,7 @@ void jsd_epd_nominal_process(jsd_t* self, uint16_t slave_id) {
 
   unsigned int controlword_uint = (unsigned int)self->slave_states[slave_id].epd_nominal.rxpdo.controlword;
   WARNING("Bits of rxpdo controlword:\n");
-  for(int bit=0;bit<(sizeof(unsigned int) * 8); bit++)
+  for(long unsigned int bit=0;bit<(sizeof(unsigned int) * 8); bit++)
   {
     WARNING("%i ", controlword_uint & 0x01);
     controlword_uint = controlword_uint >> 1;
